2nd_commit
error jwt 3 parts

+++ Avocat_Spring-master/src/main/java/com/example/gestion_user/repositories/ContributorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ContributorRepository extends JpaRepository<Contributor, Integer> {
+public interface ContributorRepository extends JpaRepository<Contributor, Long> {
     List<Contributor> findByAuxiliary(Auxiliary auxiliary);
 }
