+++ Avocat_Spring-master/src/main/java/com/example/gestion_user/repositories/CourtRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourtRepository extends JpaRepository<Court, Integer> {
+public interface CourtRepository extends JpaRepository<Court, Long> {
 }
